Task50225

See merge request virtus-hp-ML-2019/ml-git!93

+++ src/mlgit/metadata.py
@@ -141,8 +141,10 @@ class Metadata(MetadataManager):
 		#saves README.md if any
 		readme = "README.md"
 		src_readme = os.path.join(idxpath, readme)
-		dst_readme = os.path.join(fullmetadatapath, readme)
-		shutil.copy2(src_readme, dst_readme)
+
+		if os.path.exists(src_readme):
+			dst_readme = os.path.join(fullmetadatapath, readme)
+			shutil.copy2(src_readme, dst_readme)
 
 		#saves metadata and commit
 		metadata[self.__repotype]["manifest"]["files"] = "MANIFEST.yaml"
